split config and replacement
