+++ core/config.go
@@ -1,14 +1,9 @@
 package core
 
 import (
-	"bytes"
 	"os"
-	"regexp"
-	"text/template"
 
 	"github.com/Masterminds/semver/v3"
-	"github.com/icholy/replace"
-	"golang.org/x/text/transform"
 	"gopkg.in/yaml.v2"
 
 	"github.com/miniscruff/changie/shared"
@@ -20,49 +15,6 @@ const (
 
 // GetVersions will return, in semver sorted order, all released versions
 type GetVersions func(shared.ReadDirer, Config) ([]*semver.Version, error)
-
-type ReplaceData struct {
-	Version         string
-	VersionNoPrefix string
-}
-
-type Replacement struct {
-	Path    string
-	Find    string
-	Replace string
-}
-
-func (r Replacement) Execute(
-	readFile shared.ReadFiler,
-	writeFile shared.WriteFiler,
-	data ReplaceData,
-) error {
-	templ, err := template.New("replacement").Parse(r.Replace)
-	if err != nil {
-		return err
-	}
-
-	var buf bytes.Buffer
-
-	//nolint:errcheck
-	templ.Execute(&buf, data)
-
-	fileData, err := readFile(r.Path)
-	if err != nil {
-		return err
-	}
-
-	transformer := replace.Regexp(regexp.MustCompile(r.Find), buf.Bytes())
-
-	newData, _, _ := transform.Bytes(transformer, fileData)
-
-	err = writeFile(r.Path, newData, 0644)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
 
 // Config handles configuration for a changie project
 type Config struct {
